Davidl 64bit eventnumber
This contains a very large number of changes to support moving to 64bit event numbers. The signature of both the brun and evnt methods were changed in JANA 0.7.4 so this is required to be compatible with that. We should still be able to compile with JANA 0.7.3 even with these changes ( but not JANA 0.7.2 or earlier.)

+++ src/libraries/ANALYSIS/DKinFitResults_factory.cc
@@ -18,7 +18,7 @@ jerror_t DKinFitResults_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t DKinFitResults_factory::brun(jana::JEventLoop* locEventLoop, int runnumber)
+jerror_t DKinFitResults_factory::brun(jana::JEventLoop* locEventLoop, int32_t runnumber)
 {
 	vector<const DAnalysisUtilities*> locAnalysisUtilitiesVector;
 	locEventLoop->Get(locAnalysisUtilitiesVector);
@@ -96,7 +96,7 @@ void DKinFitResults_factory::Reset_NewEvent(void)
 //------------------
 // evnt
 //------------------
-jerror_t DKinFitResults_factory::evnt(JEventLoop* locEventLoop, int eventnumber)
+jerror_t DKinFitResults_factory::evnt(JEventLoop* locEventLoop, uint64_t eventnumber)
 {
 #ifdef VTRACE
 	VT_TRACER("DKinFitResults_factory::evnt()");
