+++ src/libraries/RF/DRFTime_factory_TAGH.cc
@@ -22,7 +22,7 @@ jerror_t DRFTime_factory_TAGH::init(void)
 //------------------
 // brun
 //------------------
-jerror_t DRFTime_factory_TAGH::brun(jana::JEventLoop *locEventLoop, int runnumber)
+jerror_t DRFTime_factory_TAGH::brun(jana::JEventLoop *locEventLoop, int32_t runnumber)
 {
 	//RF Period
 	vector<double> locBeamPeriodVector;
@@ -73,7 +73,7 @@ jerror_t DRFTime_factory_TAGH::brun(jana::JEventLoop *locEventLoop, int runnumbe
 //------------------
 // evnt
 //------------------
-jerror_t DRFTime_factory_TAGH::evnt(JEventLoop *locEventLoop, int eventnumber)
+jerror_t DRFTime_factory_TAGH::evnt(JEventLoop *locEventLoop, uint64_t eventnumber)
 {
 	//The RF Time is defined at the center of the target
 	//The time offset needed to line it up to the center of the target is absorbed into the calibration constants.
