+++ src/libraries/START_COUNTER/DSCHit_factory.cc
@@ -71,7 +71,7 @@ jerror_t DSCHit_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t DSCHit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
+jerror_t DSCHit_factory::brun(jana::JEventLoop *eventLoop, int32_t runnumber)
 {
     // Only print messages for one thread whenever run number change
     static pthread_mutex_t print_mutex = PTHREAD_MUTEX_INITIALIZER;
@@ -166,7 +166,7 @@ jerror_t DSCHit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t DSCHit_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t DSCHit_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
     /// Generate DSCHit object for each DSCDigiHit object.
     /// This is where the first set of calibration constants
