+++ src/libraries/TRACKING/DTrackCandidate_factory.h
@@ -57,8 +57,8 @@ class DTrackCandidate_factory:public JFactory<DTrackCandidate>{
    
  protected:
   virtual jerror_t init(void);
-  virtual jerror_t evnt(JEventLoop *loop, int eventnumber);	///< Invoked via JEventProcessor virtual method
-  virtual jerror_t brun(JEventLoop* eventLoop,int runnumber);
+  virtual jerror_t evnt(JEventLoop *loop, uint64_t eventnumber);	///< Invoked via JEventProcessor virtual method
+  virtual jerror_t brun(JEventLoop* eventLoop,int32_t runnumber);
   virtual jerror_t erun(void);
   virtual jerror_t fini(void);
