+++ src/plugins/Analysis/DAQ_WRD2PI/Df250PulseIntegral_factory.cc
@@ -39,7 +39,7 @@ jerror_t Df250PulseIntegral_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t Df250PulseIntegral_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
+jerror_t Df250PulseIntegral_factory::brun(jana::JEventLoop *eventLoop, int32_t runnumber)
 {
 	return NOERROR;
 }
@@ -47,7 +47,7 @@ jerror_t Df250PulseIntegral_factory::brun(jana::JEventLoop *eventLoop, int runnu
 //------------------
 // evnt
 //------------------
-jerror_t Df250PulseIntegral_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t Df250PulseIntegral_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 
 	// Code to generate factory data goes here. Add it like:
